feat: tested the program. debugged it, fixed the bugs. now it works
the README.md, JOURNAL logs, a post and theory are next

+++ my_reusable_drivers/utils/stack/stack.c
@@ -16,7 +16,7 @@ bool Stack_init(Stack_t *stack, uint32_t *buf, uint32_t size)
 
 bool Stack_push(Stack_t *stack, uint32_t element)
 {
-    if (stack->top >= stack->size - 1)
+    if (stack->top >= (int)stack->size - 1)
     {
         return false; // overflow
     }
